Fix some database commands that set up the SQL Server database for tests that don't properly inject arguments
SqlCommand does not currently support using parameters for database
names. I don't know if this changed, but directly interpolating the
command string is the recommended way to set the database name from what
I was able to research. I wasn't able to find any changes to SqlCommand
that would change this behavior, so I'm not sure why this was necessary
for me and why it was working before for others.